style: fix footer responsive style
